[native] Refactor arrow flight build in adapters script
Velox provides a function to install the Arrow library. We don’t need
to copy and paste the same code here and can re-use it.
There is an EXTRA_ARROW_OPTIONS variable that allows custom
Arrow library build options to be able to pass along that Arrow Flight
should be built.